Rename Barnepass.tsx til PassAvDineBarn for å fikse navnekollisjon med Barnepass-typen (og for bedre å matche overskriften på siden)
